move merge networks function to util, allow graceful timeout option
