+++ pkg/util/util.go
@@ -2,6 +2,7 @@ package util
 
 import (
 	"fmt"
+	"net"
 	"os"
 )
 
@@ -13,4 +14,18 @@ func EnsureEnvironmentSet(envKey string) error {
 		return fmt.Errorf("environment %s is not set", envKey)
 	}
 	return nil
+}
+
+func MergeNetworkMaps(networks []map[string]*net.IPNet) map[string]*net.IPNet {
+	// if there is only one network return it
+	if len(networks) == 1 {
+		return networks[0]
+	}
+	merged := make(map[string]*net.IPNet)
+	for _, nm := range networks {
+		for name, network := range nm {
+			merged[name] = network
+		}
+	}
+	return merged
 }
